T10886 make Differ with mutable index by introducing a lifetime to the trait
it was quite bad design to have a refcell inside index as that allows for the possibility of
mutating the index while only having a shared reference to it (obtained
from `with_index`). However, this won't persist the index to disk as it
assumes it won't be changed. This commit attempts avoiding having
interior mutability in the index structure