change `Deserialize` trait to use `&mut impl BufRead`
this change was made due to `take` not being callable on trait objects.
Although, weirdly `take` was being called on `&mut dyn BufRead` somehow? (see
Tree:deserialize in a commit). Turns out the problem was `std::io::Read`
not being in scope but am unsure of what's going on here precisely.

Will probably stick with `impl Trait` as the only reason it was was for
object safety (avoid generics) but none of the deserialize methods are
callable on trait objects anyway so this point is now moot.