refactor packs so the readers are reused
packfiles now own the readers rather than creating readers in each
recursion.
unfortunately, now doing anything with pack requires mutable access and
we really don't want odb methods to take `&mut self` as are accessed in
repo which only ever has shared access so we'd just have to move the
refcell up which is even worse as we'd be paying for refcell even when
its not needed. The current solution is to wrap packs in refcell but
there may be a better way.

remove .csproj to stop editor from trying to do stuff with ribble