only serialize index if a dirty flag has been set
now stuff is getting a little messy unfortunately, perhaps this is the
cost of optimization?
there are now a few subtle things that must be done correctly such as
setting the dirty flag. This is just done via the deref_mut impl which
seems rather dirty but it avoids the programmer having to set it
manually everywhere a change could be made

on the brightside, `bit status` is now significantly faster (at least 3x on libgit2 repo)