more refactoring
remove TreeIteratorEntry and just use BitIndexEntry.

This enum seemed like unnecssary noise and probably was more harm than benefit.
BitIndexEntry is general enough as it is a strict superset of a tree entry

We already have impls for converting two and from these two entry types
(obviously some bitindexentry fields will be meaningless when created from a tree entry but this hasn't turned out to be an issue)