remove old `build_tree` implementation
the new one seems to be a little bit slower which isn't surprising as it's
doing more work. can undo this change and keep the old "specialized"
implementation if it turns out to be an issue. It was extremely slow
before due to the issue in `tree_cache` where the call in
`create_pseudotree` to find a valid child would take a very long time