Prepare code for implementing checkout
remove worktree_tree_iter.
the main point of worktree iter is to get the stat information which we can't really do under the tree_iterator interface.
the primary benefit of tree iter is to be able to step over when hashes match.
but we will never know the tree hash of a worktree iter so this will never happen.
WorktreeIterator seems to make much more sense to only implement the
entryiterator interface