fix bug in indexentry ordering function
this doesn't actually affect the ordering of the index as the btreemap
key is just (BitPath, MergeStage) as is purely ordered off just that.

I've forgotten why stuff is ordered the way it is.

This is the output of the index file on the following repo (everything
added to index).

Moved implementation of ordering for index entries to test modules as
they're not actually used in real code which is quite misleading.

Everything seems to work as intended so I think I will leave this alone...
.
├── 5
├── empty
├── foo
│  ├── bar
│  └── bar.l
├── foo.bar
└── new

$ git ls-files -s
100644 2ef267e25bd6c6a300bb473e604b092b6a48523b 0       5
100644 e69de29bb2d1d6434b8b29ae775ad8c2e48c5391 0       foo.bar
100644 e69de29bb2d1d6434b8b29ae775ad8c2e48c5391 0       foo/bar
100644 e69de29bb2d1d6434b8b29ae775ad8c2e48c5391 0       foo/bar.l
100644 e69de29bb2d1d6434b8b29ae775ad8c2e48c5391 0       new