very minor optimizations by avoiding utf8 validation and an unnecessary assertion
All in all, bit status is almost twice as fast compared to 090eeab8550684aa0918dfbafaa9d39c1f8c4e0a.

status on lg2 from 60ms to 35ms.

This is still decently slower than git (2-3 times generally). But the
speed is now fast enough to be negligible on the vast majority of
repositories (in terms of git status).