Refactor tree diffing api
Instead of the badly named `GenericDiffer` we now have a DiffDriver
which is a name much more indicative of what it's really doing.
The diff driver is parameterized by a `Differ` (or TreeDiffer) which
is essentially a bunch of callbacks called by the driver.

The treediffer trait has been expanded to include `created_tree` and
`deleted_tree`.

Rename `WorkspaceDiff` to `WorkspaceStatus` as "diff" is becoming more
and more painfully overloaded