add a type representing a validated reference
not 100% clear this is necessary, but it was bothering me that we
couldn't really tell the difference between a reference that was
partially resolved to refs/heads/master for instance vs a reference that
was stuck on refs/heads/master due to that reference not existing yet.
The validated reference enum has a specific case for it.