serialize index using a buffered writer as otherwise it's unbelievable slow
On libgit2 500ms down to 40ms. It's due to the hash being updated a
billion times at tiny sizes and it was just so inefficient.
The optimization of using a BufferedWriter can probably be applied in
many other places.

In fact, most of the IO probably has huge room for improvement such as
Hashreader etc