add simple tests for merge
fix issue `Index::add_entry_common` where it would try to read the blob
from disk. However, it's possible for `Index::add_conflicted_entry` to
be called with an entry that has been deleted and no longer exists on
disk and `add_entry_common` would then fail when it tried to read the
entry. Instead, it should first check whether the entry oid is known and
if so, it doesn't need to do a write.