use tree_cache to speed up building tree's from index
It does so by again skipping over trees that already exist.

The implementation is not quite ideal yet as the root tree will not be skipped even if it's unchanged.
Although I guess that's not a major issue as we don't allow empty commits anyway, and even if we did
that would be a rare case.