be more vigilant and wary when resolving revisions and references
There was no check that the oid resolved from a reference/revision
pointed to a commit. Similarly, we could write non-commit oid's to ref
files which is definitely wrong.
Instead of just checking the object exists we now check that the object
exists AND is a commit.

For example, in the test for checkout we were passing a tree to
`checkout_reference` and it just worked and wrote the tree to the branch
which is rather bad.

This didn't play nicely with `bit cat-file` so just added some
not too bad workarounds for that.