add an unfortuante `ModifiedTree` variant so checkout can matchup the worktree and diff entries correctly
With an tree iterator for worktree, if no diff_entry is yielded when
there's a modified tree entry then the worktree iterator will get desynced
(as it will yield a directory entry regardless if it's changed or not) and everything will go rong.

There's probably a way around this variant as the only place that needs it is checkout atm,
but this is relatively simple and understandable for now.