Wemove `with_index` api and just use an accessor `repo.index()?`
To maintain the automatic rollback of the index we put the check in the
`repo.enter` method instead.
This makes more sense as if we had multiple `with_index` calls before it
could try to rollback multiple times, or even try to commit after a
rollback which would panic.
Now we can consider a call to `enter` to be one big transaction for the
index